ui: display recent records on community page
(closes #1699)
